dev - reorganized signup and login for firebase; implemnted basic creation of user within db on signup
