+++ repay-app/Home/styles.js
@@ -1,6 +1,6 @@
 import { StyleSheet } from 'react-native';
 
-export const styles = StyleSheet.create({
+export default StyleSheet.create({
     container: {
       //flex: 1,
     },
